third_party/WebKit/LayoutTests/platform/win/external/wpt/html/browsers/history/the-history-interface/005-expected.txt

This is a testharness.js-based test.
PASS history.pushState support is needed for this testcase 
PASS <body onpopstate="..."> should register a listener for the popstate event 
PASS window.onpopstate should register a listener for the popstate event 
Harness: the test ran to completion.
